Add files via upload
one of the output variables of the script (yp(34)) should be F_sc (flow spent catalyst) rather than F_sp (flow of catalyst in standpipe). In other words, yp(34) = F_sc rather than yp(34)=F_sp. This error does not affect in any sense any of the results of the papers (MPC, DL, PP etc.) since yp(34) is only used in a plotting script (to plot the flow rates among other variables in the open source model paper) . In other words, we do not use yp(34) nor  F_sc or F_sp in any of our frameworks. this is corrected in this version of Fcc.m